requirements.txt
requests>=2.32.0
pandas>=2.2.3
numpy>=2.1.0
scikit-learn>=1.5.0
python-dotenv>=1.0.1
PyMySQL>=1.1.0
pytz>=2024.1
